prueba3/TestDjango/requirements.txt
pillow
django-crispy-forms
cx_oracle
djangorestframework
